Separate learning in a new file.
